requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
idna==2.6
iniherit==0.3.9
mock==2.0.0
multidict==4.1.0
nose==1.3.7
pbr==3.1.1
python-slugify==1.2.4
PyYAML==3.12
requests==2.18.4
requests-mock==1.4.0
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
wrapt==1.10.11
yarl==1.1.0
yoyo-migrations==5.0.5
